[example] simplify opt example
