+++ examples/language/gpt/train_gpt_demo.py
@@ -5,10 +5,8 @@ from time import time
 import psutil
 import torch
 import torch.nn as nn
-from model_zoo import model_builder
 from packaging import version
 from torch.nn.parallel import DistributedDataParallel as DDP
-from utils import get_data, get_tflops
 
 import colossalai
 from colossalai.logging import disable_existing_loggers, get_dist_logger
@@ -16,6 +14,8 @@ from colossalai.nn.parallel import ZeroDDP
 from colossalai.tensor import ColoParameter, ComputePattern, ComputeSpec, ProcessGroup, ReplicaSpec, ShardSpec
 from colossalai.utils import get_current_device
 from colossalai.utils.model.colo_init_context import ColoInitContext
+from examples.language.commons.utils import get_data, get_tflops
+from model_zoo import model_builder
 
 CAI_VERSION = colossalai.__version__
